+++ hc/front/forms.py
@@ -125,6 +125,16 @@ class AddWebhookForm(forms.Form):
         return json.dumps(dict(self.cleaned_data), sort_keys=True)
 
 
+class AddShellForm(forms.Form):
+    error_css_class = "has-error"
+
+    cmd_down = forms.CharField(max_length=1000, required=False)
+    cmd_up = forms.CharField(max_length=1000, required=False)
+
+    def get_value(self):
+        return json.dumps(dict(self.cleaned_data), sort_keys=True)
+
+
 phone_validator = RegexValidator(
     regex="^\+\d{5,15}$", message="Invalid phone number format."
 )